requirements.txt
pandas 
requests 
beautifulsoup4 
tqdm
sentence-transformers 
faiss-cpu
fastapi 
uvicorn 
streamlit
nltk